+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
-## v3.0.0-rc.??? (2021-06-??)
+## v3.0.0-rc.205 (2021-06-24)
 
 #### :rocket: New Feature